% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\docType{data}
\name{environment}
\alias{environment}
\title{Three fictive environmental variables
A dataset containing the values of three fictive envionemental variables
in 100 samples. The corresponding pseudo-species abundance in each sample
are available in the data_abundance dataset.}
\format{
A data frame with 100 rows and 3 variables:
\describe{
\item{x1}{an environmental gradient from 0 to 25}
\item{x2}{an environmental gradient from 0 to 40}
\item{x3}{an environmental gradient from 1 to 0}
...
}
}
\usage{
data(environment)
}
\description{
Three fictive environmental variables
A dataset containing the values of three fictive envionemental variables
in 100 samples. The corresponding pseudo-species abundance in each sample
are available in the data_abundance dataset.
}
\author{
Loïck Kléparski \email{loick.kleparski@hotmail.fr}
}
\keyword{data}
