% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanmean4.R
\name{nanmean4}
\alias{nanmean4}
\title{Mean of the n highest values of a vector
Estimates the mean of the n\% of the samples in x with the highest abundance}
\usage{
nanmean4(x, n)
}
\arguments{
\item{x}{a vector with the values to average}

\item{n}{an integer corresponding to the percentage of values to use}
}
\value{
mean of the n\% of the samples in x with the highest values
}
\description{
Mean of the n highest values of a vector
Estimates the mean of the n\% of the samples in x with the highest abundance
}
