% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_difer2.R
\name{niche_difer2}
\alias{niche_difer2}
\title{Nested function for niche overlapping estimation with 'combina_niche3.R' and 'niche_differ_sp.R'.}
\usage{
niche_difer2(sp_chr1, sp_chr2, T)
}
\arguments{
\item{sp_chr1}{a matrix (niche categories by environmental variables) for species 1}

\item{sp_chr2}{a matrix (niche categories by environmental variables) for species 2}

\item{T}{an integer corresponding to the threshold of minimal abundance in a category for the niche breadth estimation}
}
\value{
index D
}
\description{
Estimate the niche overlapping (index D) between two species. Assume that the niche is rectangular.
Columns (i.e. environmental variables) in 'spe_chr1' and 'sp_chr2' have to be the same.
}
