% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_difer_sp.R
\name{niche_difer_sp}
\alias{niche_difer_sp}
\title{Nested function for niche overlapping estimations with 'combina_niche3.R'}
\usage{
niche_difer_sp(sp_chr, T)
}
\arguments{
\item{sp_chr}{a three dimensional matrices with the species chromatograms (alpha category by p environmental variables by species)}

\item{T}{an integer corresponding to the threshold of minimal abundance in a category for the niche breadth estimation}
}
\value{
index D
}
\description{
Nested function for niche overlapping estimations with 'combina_niche3.R'
}
