% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti_eury_niche2_v2.R
\name{opti_eury_niche2_v2}
\alias{opti_eury_niche2_v2}
\title{Optimums and amplitudes estimations along each niche dimension}
\usage{
opti_eury_niche2_v2(sp_chr, T, z, y, k)
}
\arguments{
\item{sp_chr}{a matrix with the species chromatograms (categories by environmental variables by species). Outputs of 'chromato_env16.R'}

\item{T}{an integer corresponding to the min abundance threshold for niche amplitudes estimations}

\item{z}{a matrix with n samples by p environmental variables (i.e. the value of each environmental varibale in each sample)}

\item{y}{a matrix with the species abundance in the n samples}

\item{k}{an integer corresponding to the percentage of samples with the highest abundance values to use to estimate the mean abundance in a given category}
}
\value{
deg_eury a matrix with the degree of euryecie (niche breadth) for each species along each environmental dimensions (matrix environmental variable by species)

mean_deg_eury a matrix with the mean degree of euryecie of each species

opti_val a matrix with the niche optimum values (matrix environmental variables by species)
}
\description{
Optimums and amplitudes estimations along each niche dimension
}
