% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opti_eury_niche2.R
\name{opti_eury_niche2}
\alias{opti_eury_niche2}
\title{Optimums and amplitudes estimations along each niche dimension.}
\usage{
opti_eury_niche2(sp_chr, Thres_T, z, y, k)
}
\arguments{
\item{sp_chr}{a matrix with the species chromatograms (categories by environmental variables by species). Outputs of 'chromato_env16.R'}

\item{Thres_T}{an integer corresponding to the min abundance threshold for niche amplitudes estimations}

\item{z}{a matrix with n samples by p environmental variables (i.e. the value of each environmental variable in each sample). Same matrix as in 'chromato_env16.R'.}

\item{y}{a matrix with the species abundance in the n samples}

\item{k}{an integer corresponding to the percentage of samples with the highest abundance values to use to estimate the mean abundance in a given category. Should have the same value as in 'chromato_env16.R'}
}
\value{
Three matrices are returned:

amplitudes, a matrix with the degree of euryoecie (niche breadth) of each species (in column) along each environmental dimension (in line)

mean_amplitudes, a matrix with the mean degree of euryoecie of each species

optimums, a matrix with the niche optimum values of each species (in column) along each environmental dimension (in line)
}
\description{
Function that estimates niche optimums and amplitudes along each environmental variable and for each species.
}
